+++ mediation/UnityAds/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Unity Ads Adapter plugin for Google Mobile Ads SDK for Unity 3D Changelog
 
+## 1.1.2
+- Supports Unity Ads Android SDK version 2.2.0.
+- Supports Unity Ads iOS SDK version 2.2.0.
+- Added `UnityAds.SetGdprConsentMetaData()` method to manually request a user opt-in for targeted advertising.
+
+## 1.1.1
+- Supports Unity Ads Android SDK version 2.2.0.
+- Supports Unity Ads iOS SDK version 2.2.1.
+
 ## 1.1.0
 - Supports Unity Ads Android SDK version 2.2.0.
 - Supports Unity Ads iOS SDK version 2.2.0.